Allow filter/replace on IDs without a description.
See #47; partially reverts 515f55eb7ed2607b3c216ca5c8ced4bae8df51a5.

The sticky part is replacement. This commit's approach:

Substitution is always performed on the sequence ID. If the first "word"
of the description matches the ID, assume that the description is
FASTA-like and perform substitution only on the "rest" of the
description, then add the (possibly modified) ID on front. If the first
word does not match the ID, perform substitution on the entire
description.