Rework `--pattern-replace` logic per @cmccoy.
See #48. If the ID and the first word of the description match, apply
the transform to the entire description and then set the ID from the
first word. If the ID and the first word of the description do not
match, apply the transform to each individually.

Also added a few real-world test cases from StackOverflow.