Fix for Biopython 1.71 dual coding support.
This closes issue #73, indirectly triggered by changes
in Biopython to support recent NCBI codon tables with
codons which can be amino acids of stop codons.

Previously the monkey-patched sub-class was breaking here:

dual_coding = [c for c in stop_codons if c in forward_table]

This change adds direct support for __contains__ by
forwarding this to the wrapped forward table.